Backchanges from windows version
git-svn-id: http://calder.sdml.cs.kent.edu/svn/srcML/trunk@2727 f70b1346-9fe7-0310-9bf3-95a0de396e6b

+++ src/StateStack.h
@@ -136,7 +136,7 @@ class StateStack {
     if (st.empty())
       throw Segmentation_Fault();
 
-    st.top().push(id);
+    st.top().push((int) id);
   }
 
   void pop() {
